Making site work with Sites framework.
Some changes I meant to commit in the last one, relating to
tidying up the Aggregator so that it relates to the Sites framework
better, and doesn't duplicate as much code.

And new stuff that should make the Blogs and Entries work with the
Sites framework, so that multiple Sites can be created (each Site
will need its own Aggregator too).