Add Place/Country model.
We now have a separate model for Places, with a sub-class of Country. These should
be easy to extend to a more full GeoPlanet-enabled system in the future, and use
WOE IDs at the moment.

We can now assign one or more Countries to a Movie.
And can assign one Country to Cinema.