Make main body text bigger.
Based on line-height of 36px rather than 24px, so going up to 20px
font-size.

Then tweak the widths of the main-col-only column at certain widths
to look better with the new font size.

And then fix the width of .figure-full elements at wider widths
given the new column widths (and use a mixin for it).