Tweak CSS for meta blocks
Retain the alighment of left border with the top and bottom of the text
but move the `float:left` to the `.meta__inner`, and apply clearfix to
the `.meta` block.

Because, when we were floating the `.meta` block, elements following it
would collapse up into it, losing their top margin.

For #99