Fix spacing of headers
We'd made a tweak to add spacing above headers, because that fixed
the "Comments" header on post_detail.html and <h*> headers that
appear in blog posts immediately after <hr> tags.

But that screwed up all the headers within .media and .gallery items
(and maybe others).

So revert those changes and add different, more focused fixes for
the "Comments" header (a new utility class for just adding some
margin-top) and headers after <hr> tags.

For #99