Add consistent "comments are closed" messages across all template sets
It won't work on the older template sets if we turn all comments on,
because they just have their old comment forms disabled, so they
(a) won't have a form and (b) won't have a message. But that's no
worse than the current situation.

At least now, if we turn off comments for posts older than 30 days (say)
then they'll have a message showing that.

We could set all the Posts older than the latest template set to have
comments turned off, which would then always show a message?

For #99