Add content type header to RSS feeds
The "correct" content type of "application/rss+xml" means that some
browsers will download the RSS feed, or try to open it in an app.

But we want to display it in the browser, styled with the XSLT
stylesheet. Changing the content type to "application/xml" allows this,
and it's what a bunch of places use for their feeds' content type
anyway.

For #248