Fix static files tests
Turned out that they only worked if the CSS/JS files hadn't changed/moved since the last time `collectstatic` was run. So, to test the file exists by doing a `.get()` on it, we'd have had to run that before running the test(s), which slows things up a lot. This method ensures it exists, regardless of `collectstatic`.