Add redirect URL for individual Tweets
When clicking the VIEW ON SITE link in the Django Admin edit view,
we got a 500 error, because we don't include the Django Ditto Twitter
URLs, which `Tweet.get_absolute_url()` needs.

So we've added a URL named `twitter:tweet_detail` which redirects to
the correct `DayArchiveView` for the tweet in question.

And added an #anchor to each tweet in that page's template so we
can link directly to one of them.

Fixes #396