Replace RSS SVG with a smaller one
The `yarn build` step in Dockerfile was stalling for no apparent reason.
After a lot of trial and error, it turned out to be the RSS SVG. If I
removed some of it, things worked. I assume that it was too long, but
no idea why it used to work.

So I've removed the gradient background, replacing with a solid colour.