In Webmentions Admin RSS feed ignore webmentions that aren't to Posts
Turns out that we always accept webmentions to ANY URL, not just those
that point to a model with MentionableMixin. But those extra webmentions
won't have a GenericForeignKey relation to anything, and our RSS
feed code was expecting that.

Seeing as we don't display or invite mentions on any pages that aren't Posts,
we're going to just ignore those extra webmentions in the RSS feed.